LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/canvas/element/drawing-text-to-the-canvas/2d.text.measure.emHeights-expected.txt

2d.text.measure.emHeights
Testing emHeights
Actual output:

FAIL Testing emHeights assert_equals: ctx.measureText('A').emHeightAscent === 37.5 (got 86[number], expected 37.5[number]) expected 37.5 but got 86
